+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - content changes to articles of association task in conversions and transfers
   to describe the process more accurately
 - addition of 1 checkbox to articles tasks in conversions and transfers
+- the all project by trust view now includes 'form a multi academy trusts'
 
 ## [Release-80][release-80]